Added node and links animations. Added the ability to forever highlight links. Added alignment for the nodes with user choice between Left, Right, Justified.
Modified the positions a bit to make things look prettier.

Added a piece of code, that is supposed to run python scripts in browser. The code works, but the XSLX file that is parsed as JSON to the python script has issues.
Separately, I added the ability for the user to choose a dataset (that follows the standards) and have it be processed into a JSON file with defined links, nodes and values. This is WIP as it has some issues with the way it generates links.